kn;;[pk

+++ Server/Routes/admin.js
@@ -2,7 +2,8 @@ const express  = require("express")
 const router  = express.Router();         
 const BLOG_DB = require("../Schemas/blogSchema");
 const USERS = require("../Schemas/userSchema")
-const{upload , saveImageToS3 ,deleteImageFromS3, randomBytes,sendEmail,sendEmails} = require("../Utils/adminUtils");
+const SOME_EMAILS = require("../Schemas/newsLetterSchema")
+const{upload , saveImageToS3 ,deleteImageFromS3, randomBytes,sendEmails,isAnAdmin} = require("../Utils/adminUtils");
 
 // router.use(express.json());
 // router.use(express.urlencoded({extended:false}));
@@ -26,7 +27,9 @@ router.post("/post", upload.single("image") ,async function(req,res){
 
 router.post("/newsletter" , async function(request , response){
     console.log(request.body);
-    const recipients = await USERS.find().where("newsLetter").equals(true).select("email");
+    const first_Array_of_emails = await USERS.find().where("newsLetter").equals(true).select("email");
+    const second_Array_of_emails = await SOME_EMAILS.find().select("email");
+    const recipients = [...first_Array_of_emails, ...second_Array_of_emails];
     try{
        await sendEmails(recipients, request.body.subject, request.body.textBody );
        response.send("email sent");
@@ -37,11 +40,15 @@ router.post("/newsletter" , async function(request , response){
 
 
 router.delete("/delete/:title", async function(req, res){
-   const blogTitle = req.params.title;
-   const blogImageName = await BLOG_DB.find().where("title").equals(blogTitle).select("image");
-   await deleteImageFromS3(blogImageName[0].image);
-   await BLOG_DB.deleteOne({title:blogTitle});
-   res.send("blog post deleted");
+    if(req.user["0"].name !== undefined && isAnAdmin(req.user["0"].name)){
+         const blogTitle = req.params.title;
+         const blogImageName = await BLOG_DB.find().where("title").equals(blogTitle).select("image");
+         await deleteImageFromS3(blogImageName[0].image);
+         await BLOG_DB.deleteOne({title:blogTitle});
+         res.send("blog post deleted");
+    }
+    
+    res.send("cheeky corny bastard")
 })
